+++ app/frontend/src/javascript/api/subscription.ts
@@ -1,10 +1,10 @@
 import apiClient from './clients/api-client';
-import { Subscription, SubscriptionPaymentDetails, UpdateSubscriptionRequest } from '../models/subscription';
+import { Subscription, SubscriptionPaymentDetails } from '../models/subscription';
 import { AxiosResponse } from 'axios';
 
 export default class SubscriptionAPI {
-  static async update (request: UpdateSubscriptionRequest): Promise<Subscription> {
-    const res: AxiosResponse<Subscription> = await apiClient.patch(`/api/subscriptions/${request.id}`, { subscription: request });
+  static async cancel (id: number): Promise<Subscription> {
+    const res: AxiosResponse<Subscription> = await apiClient.patch(`/api/subscriptions/${id}/cancel`);
     return res?.data;
   }